Nickcoin
Nickcoin is an experimental digital currency that enables instant payments to anyone, anywhere in the world. Nickcoin uses peer-to-peer technology to operate with no central authority: managing transactions and issuing money are carried out collectively by the network. Nickcoin Core is the name of open source software which enables the use of this currency.

For more information, as well as an immediately useable, binary version of the Nickcoin Core software